Run locally to debug when you actually want to!
No more forgetting to switch this variable! It does mean that when 
xdebug is intended to be used, drush *will always* run locally, so 
beware. But this is usually the intention anyway, as leaving the 
XDEBUG_CONFIG environment variable set will slow things down otherwise.

Note this also includes a tweak to my ZSH theme, as it turns out you can
set (export) the XDEBUG_CONFIG variable to an empty string / null, and
XDebug/PHPStorm will still consider that enough to catch debugging. So
reflect that in the debugging indicator in the prompt.